Consolidate catalog types and split feature oracle
